requirements.txt
anthropic>=0.18.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
streamlit==1.39.0
pandas==2.2.0
chromadb==0.4.22
sentence-transformers==2.3.1
